==> Address/prajwal/Feb 20/bk 1850/address 2/lib/05_combine-new3.ipynb <==
import pandas as pd
#from fuzzywuzzy import fuzz
import pickle

#mn_or_bk = 'mn' #bk or mn
#year = str(1850) #1850, 1880
#doubttype = 'doubt' #doubt or not so confident

#df1 = pd.read_csv(mn_or_bk + '_' + year + '/S/' + doubttype + '-' + mn_or_bk + year + '-v7-s.csv')
#df1 = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run1/mn-1880-3-s.csv')
df1 = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data2/bk-1850-3-s-address2.csv')


df1.head()

#with open('/Users/prajwal/Desktop/Columbia/C4SR/' + mn_or_bk + '_' + year + '/' + doubttype + '-' + mn_or_bk + year + '-v7.pkl', 'rb') as f:
  #d = pickle.load(f)
with open('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data/bk-1850-3-address2.pkl', 'rb') as f:
    d = pickle.load(f)

#df1_original = pd.read_excel(mn_or_bk + '_' + year + '/' + doubttype + '-' + mn_or_bk + year + '-v7.xlsx')



df1_original = pd.read_excel('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data/bk_1850_address2_afterDoubt.xlsx')


len(df1)-len(df1_original)

#df1_final_1 = df1_original.loc[df1_original['filter11']==0][['n_street1','n_street2','c_street1','c_street2']]
df1_final_1 = df1_original[['n_street1','n_street2','c_street1','c_street2']]

len(df1_final_1)

#df1_original

#df1_final_1

#df1

df1.columns

df1_final_1['Full Address'] = df1['Full Address']



df1_final_1['H Address'] = df1['H Address']




df1_final_1['H House Number'] = df1['H House Number']
df1_final_1['H Street Name'] = df1['H Street Name']
#df1_final_1['new h house number'] = df1['new h house number']
#df1_final_1['new h street name'] = df1['new h street name']
#df1_final_1['new h city'] = df1['new h city']



df1_final_1['H City'] = df1['H City']
df1_final_1['Address 2'] = df1['Address 2']
df1_final_1['Address 2 House Number'] = df1['Address 2 House Number']
df1_final_1['Address 2 Street Name'] = df1['Address 2 Street Name']
df1_final_1['Address 2 City'] = df1['Address 2 City']



#df1_final_1['unidentified h info'] = df1['unidentified h info']

df1_final_1

#df1_original[['Full Address','H House Number','H Street Name','new h house number','new h street name','new h city']]
#df1[['unidentified h info']]

df1_final_1.columns

#df1_final_1.to_csv(mn_or_bk + '_' + year + '/' + mn_or_bk + '_' + year + '_combine-' + doubttype + '4.csv')
#df1_final_1.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-3-s-combine.csv')
df1_final_1.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/bk 1850/address 2/data2/bk-1850-3-s-address2-combine.csv')


## Column descriptions:

#### Full Address
contains the cleaned entry for the address from the input file
#### H House Number
contains the house number if present in Full Address
#### H Street Name
contains the street name if present in Full Address
#### c_street1
contains the first street name in case the address was located at a corner
#### c_street2
contains the second street name in case the address was located at a corner
#### n_street1
contains the first street name in case the address was located near another street
#### n_street2
contains the second street name in case the address was located near another street
#### new h city
contains the city name (if present) in Full Address
#### new h house number
contains the house number if present in Full Address for some additional entries that are not in 'H House Number' 
#### new h street name 
contains the street name if present in Full Address for some additional entries that are not in 'H Street Name'
